backend/requirements.txt
fastapi==0.115.3
uvicorn==0.32.0
pydantic==2.9.2
langchain==0.3.4
langchain-community==0.3.3
langchain-ollama==0.2.0
python-dotenv==1.0.1
playwright==1.48.0
requests==2.32.3
aiohttp==3.10.10

